backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
pyodbc==5.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
python-dotenv==1.0.1
paramiko==3.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
xlwt==1.3.0
xlrd==2.0.1